Node system - Added numeric struct and utilized it in nodes

+++ src/Artemis.VisualScripting/Nodes/Color/HslSKColorNode.cs
@@ -3,20 +3,20 @@ using SkiaSharp;
 
 namespace Artemis.VisualScripting.Nodes.Color
 {
-    [Node("HSL Color", "Creates a color from hue, saturation and lightness values", "Color", InputType = typeof(float), OutputType = typeof(SKColor))]
+    [Node("HSL Color", "Creates a color from hue, saturation and lightness values", "Color", InputType = typeof(Numeric), OutputType = typeof(SKColor))]
     public class HslSKColorNode : Node
     {
         public HslSKColorNode() : base("HSL Color", "Creates a color from hue, saturation and lightness values")
         {
-            H = CreateInputPin<float>("H");
-            S = CreateInputPin<float>("S");
-            L = CreateInputPin<float>("L");
+            H = CreateInputPin<Numeric>("H");
+            S = CreateInputPin<Numeric>("S");
+            L = CreateInputPin<Numeric>("L");
             Output = CreateOutputPin<SKColor>();
         }
 
-        public InputPin<float> H { get; set; }
-        public InputPin<float> S { get; set; }
-        public InputPin<float> L { get; set; }
+        public InputPin<Numeric> H { get; set; }
+        public InputPin<Numeric> S { get; set; }
+        public InputPin<Numeric> L { get; set; }
         public OutputPin<SKColor> Output { get; }
 
         #region Overrides of Node
